Crate relay_pii

Scrubbing of personally identifiable information (PII) from events

Modules

transform

A transforming serde Deserializer`.

Structs

AliasRule

An alias for another rule.

CompiledPiiConfig

A representation of PiiConfig that is more (CPU-)efficient for use in PiiProcessor.

DataScrubbingConfig

Configuration for Sentry’s datascrubbing

JsonScrubVisitor

Visitor for JSON file scrubbing. It will be used to walk through the structure and scrub PII based on the config defined in the processor.

LazyPattern

Wrapper for the regex and the raw pattern string.

MultipleRule

A rule that dispatches to multiple other rules.

PatternRule

A rule that matches a regex pattern.

PiiAttachmentsProcessor

A PII processor for attachment files.

PiiConfig

A set of named rule configurations.

PiiProcessor

A processor that performs PII stripping.

RedactPairRule

A pair redaction rule.

ReplaceRedaction

Replaces a value with a specific string.

RuleSpec

A single rule configuration.

Vars

Configuration for rule parameters.

Enums

InvalidSelectorError

Error for invalid PII selectors.

JsonScrubError

Describes the error cases that can happen during JSON scrubbing.

PiiConfigError

An error returned when parsing PiiConfig.

Redaction

Defines how replacements happen.

RuleType

Supported scrubbing rules.

ScrubEncodings

Which encodings to scrub for scrub_bytes.

ScrubMinidumpError

An error returned from PiiAttachmentsProcessor::scrub_minidump.

SelectorPathItem

A path component in a composit SelectorSpec.

SelectorSpec

A selector that can match paths of processing states.

Functions

scrub_graphql

Scrubs GraphQL variables from the event.

selector_suggestions_from_value

Walk through a value and collect selectors that can be applied to it in a PII config. This function is used in the UI to provide auto-completion of selectors.